hostname: Support IPv6
Rely on netaddr for checking IP addresses syntax. This isn't a big burden,
as it is already indirectly installed on the engine machine.

Rely on 'ip addr' output order and use the interface name from the
interface lines. Before that, we used the one in the 'inet' line, but
'inet6' lines do not have it.

Test if an FQDN that resolves only to an IPv6 address is considered
as resolved by DNS.

Make 'dig' return any type of result, instead of the default (which
claims to be 'A' - IPv4 address records - but actually also includes
CNAME), but not return any of the other stuff it returns by default
(comments etc.).

Change-Id: I4667c96b173308a429ce5992d5478dc47d500984
Bug-Url: https://bugzilla.redhat.com/1295041
Signed-off-by: Yedidyah Bar David <didi@redhat.com>